Got time based attacking working.
git-svn-id: svn+ssh://fileserverness.ath.cx/serv/svn/virus-td@33 05766cc9-4f33-4ba7-801d-bd015708efd9

+++ BasicUnit.h
@@ -9,6 +9,6 @@ public:
 	//BasicUnit(void);
 	~BasicUnit(void);
 	void draw();
-	void step(float dt);
+	void step(int dt);
    unit_dir move(unit_dir direction);
 };
